Extract SortOptions into its own component
