+++ src/frontend/components/reviews.js
@@ -7,8 +7,6 @@ import Tooltip from '@reach/tooltip';
 
 // Material UI imports
 import { makeStyles } from '@material-ui/core/styles';
-import ArrowDownwardIcon from '@material-ui/icons/ArrowDownward';
-import ArrowUpwardIcon from '@material-ui/icons/ArrowUpward';
 import Box from '@material-ui/core/Box';
 import Chip from '@material-ui/core/Chip';
 import Container from '@material-ui/core/Container';
@@ -54,7 +52,7 @@ import NotFound from './not-found';
 import PreprintCard from './preprint-card';
 import PrivateRoute from './private-route';
 import SearchBar from './search-bar';
-//import SortOptions from './sort-options';
+import SortOptions from './sort-options';
 import WelcomeModal from './welcome-modal';
 
 // constants
@@ -563,90 +561,3 @@ export default function Reviews() {
     );
   }
 }
-
-function SortOptions({ sort, order, onChange }) {
-  const classes = useStyles();
-  const isMobile = useIsMobile();
-
-  return (
-    <Grid container item className={classes.sortOptions} alignItems="center">
-      {['recentRapid', 'recentFull', 'recentRequests', 'datePosted'].map(
-        name => (
-          <Grid container item xs key={name}>
-            <Box display="none">
-              <Input
-                type="radio"
-                id={`sort-options-${name}`}
-                name={name}
-                value={name}
-                display="none"
-                onClick={e => {
-                  const sortOption = e.target.value;
-                  onChange(sortOption, order === 'asc' ? 'desc' : 'asc');
-                }}
-              />
-            </Box>
-            <Grid item>
-              <Tooltip
-                label={`Sort by ${
-                  name === 'recentRapid'
-                    ? 'Date of latest Rapid Review'
-                    : name === 'recentFull'
-                    ? 'Date of latest Full Review'
-                    : name === 'recentRequests'
-                    ? 'Date of latest request for review'
-                    : 'date posted on preprint server'
-                }`}
-              >
-                <InputLabel
-                  htmlFor={`sort-options-${name}`}
-                  cursor={sort === name ? 'default' : 'pointer'}
-                  focus={sort === name}
-                >
-                  {name === 'recentRapid'
-                    ? isMobile
-                      ? 'Rapid Reviewed'
-                      : 'Recently Rapid Reviewed'
-                    : name === 'recentFull'
-                    ? isMobile
-                      ? 'Reviewed'
-                      : 'Recently Reviewed'
-                    : name === 'recentRequests'
-                    ? isMobile
-                      ? 'Requested'
-                      : 'Recently Requested'
-                    : isMobile
-                    ? 'Published'
-                    : 'Date Published'}
-                </InputLabel>
-              </Tooltip>
-            </Grid>
-            <Grid item>
-              {order === 'asc' ? (
-                <ArrowUpwardIcon
-                  visibility={sort === name ? 'visible' : 'hidden'}
-                />
-              ) : (
-                <ArrowDownwardIcon
-                  visibility={sort === name ? 'visible' : 'hidden'}
-                />
-              )}
-            </Grid>
-          </Grid>
-        ),
-      )}
-    </Grid>
-  );
-}
-
-SortOptions.propTypes = {
-  onChange: PropTypes.func.isRequired,
-  sort: PropTypes.oneOf([
-    'recentRapid',
-    'recentFull',
-    'recentRequests',
-    'datePosted',
-    '',
-  ]),
-  order: PropTypes.string,
-};
